+++ tests/integrity_checker_test.py
@@ -1,5 +1,5 @@
 import ast
-from unittest import TestCase, skip
+from unittest import TestCase
 
 from darglint.integrity_checker import IntegrityChecker
 from darglint.darglint import get_function_descriptions
@@ -162,20 +162,114 @@ class IntegrityCheckerTestCase(TestCase):
         self.assertTrue(isinstance(error, ExcessRaiseError))
         self.assertEqual(error.name, 'ZeroDivisionError')
 
-    @skip('We are going to change how everything is parsed.')
+    def has_no_errors(self, program):
+        tree = ast.parse(program)
+        functions = get_function_descriptions(tree)
+        checker = IntegrityChecker()
+        checker.run_checks(functions[0])
+        self.assertEqual(len(checker.errors), 0)
+
     def test_noqa_after_excess_raises(self):
         program = '\n'.join([
             'def some_function():',
             '    """Raise an error.',
             '',
             '    Raises:',
-            '        Exception: In all cases.  # noqa: F402',
+            '        Exception: In all cases.  # noqa: I402',
             '',
             '    """',
             '    pass',
         ])
-        tree = ast.parse(program)
-        functions = get_function_descriptions(tree)
-        checker = IntegrityChecker()
-        checker.run_checks(functions[0])
-        self.assertEqual(len(checker.errors), 0)
+        self.has_no_errors(program)
+
+    def test_noqa_for_missing_raises(self):
+        program = '\n'.join([
+            'def some_function():',
+            '    """No problems.',
+            '',
+            '    # noqa: I401 Exception',
+            '',
+            '    """',
+            '    raise Exception("No, actually there are problems.")',
+        ])
+        self.has_no_errors(program)
+
+    def test_noqa_for_excess_parameters(self):
+        program = '\n'.join([
+            'def excess_arguments():',
+            '    """Excess arguments.',
+            '',
+            '    Args:',
+            '        x: Will be here eventually.  # noqa: I102',
+            '',
+            '    """',
+            '    pass'
+        ])
+        self.has_no_errors(program)
+
+    def test_noqa_for_missing_parameters(self):
+        program = '\n'.join([
+            'def function_with_missing_parameter(x, y):',
+            '    """We\'re missing a description of x, y.',
+            '',
+            '    # noqa: I101',
+            '',
+            '    """',
+            '    print(x / 2)',
+        ])
+        self.has_no_errors(program)
+
+    def test_noqa_missing_return_parameter_added(self):
+        program = '\n'.join([
+            'def function_without_return():',
+            '    """This should have a return in the docstring.',
+            '',
+            '    # noqa: I201',
+            '',
+            '    """',
+            '    global bad_number',
+            '    return bad_number',
+        ])
+        self.has_no_errors(program)
+
+    def test_noqa_excess_return(self):
+        program = '\n'.join([
+            'def will_be_defined_later():',
+            '    """Return will be defined later.',
+            '',
+            '    Returns:',
+            '        Some value yet to be determined.',
+            '',
+            '    # noqa: I202',
+            '',
+            '    """',
+            '    pass',
+        ])
+        self.has_no_errors(program)
+
+    def test_noqa_for_missing_yield(self):
+        program = '\n'.join([
+            'def funtion_with_yield():',
+            '    """This should have a yields section.',
+            '',
+            '    # noqa: I301',
+            '',
+            '    """',
+            '    yield 3',
+        ])
+        self.has_no_errors(program)
+
+    def test_noqa_for_excess_yield(self):
+        program = '\n'.join([
+            'def function_with_yield():',
+            '    """This should not have a yields section.',
+            '',
+            '    Yields:',
+            '        A number.',
+            '',
+            '    # noqa: I302',
+            '',
+            '    """',
+            '    print(\'Doesnt yield\')',
+        ])
+        self.has_no_errors(program)